backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
sqlalchemy==2.0.20
python-multipart==0.0.6
python-jose==3.3.0
langchain
langchain-community
langchain-openai
tiktoken
openai